+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## 0.2.0
+ - Feature: Write all x-forwarded-for IPs as array in new field
+   and detect invalid IPs, tagging the event if necessary
+
 ## 0.1.1
  - fix: Fix edge case where x-forwarded-for is empty or only contains trusted IPs